Add certificate download option

+++ pki-app/src/main/java/bsep/pkiapp/repository/CertificateChainRepository.java
@@ -1,4 +1,5 @@
 package bsep.pkiapp.repository;
+import java.math.BigInteger;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -8,6 +9,6 @@ import bsep.pkiapp.model.CertificateChain;
 @Repository
 public interface CertificateChainRepository extends JpaRepository<CertificateChain, Long> {
 
-	public CertificateChain getCertificateChainBySignerSerialNumber(Long signerSerialNumber);
+	public CertificateChain getCertificateChainBySignerSerialNumber(BigInteger signerSerialNumber);
 
 }
